Refactor Trash Bin page to use controller and improve structure
